acpi: fadt: avoid warnings on 64 bit build.

+++ src/acpi/fadt/fadt.c
@@ -81,9 +81,11 @@ static int fadt_test1(fwts_framework *fw)
 		/*  Sanity check sizes with extended address variants */
 		fwts_log_info(fw, "FADT is greater than ACPI version 1.0");
 		if ((uint64_t)port != fadt->x_pm1a_cnt_blk.address) 
-			fwts_failed_medium(fw, "32 and 64 bit versions of FADT pm1_cnt address do not match (0x%8.8x vs 0x%16.16llx).", port, fadt->x_pm1a_cnt_blk.address);
+			fwts_failed_medium(fw, "32 and 64 bit versions of FADT pm1_cnt address do not match (0x%8.8x vs 0x%16.16llx).",
+				port, (unsigned long long int)fadt->x_pm1a_cnt_blk.address);
 		if (width != fadt->x_pm1a_cnt_blk.register_bit_width)
-			fwts_failed_medium(fw, "32 and 64 bit versions of FADT pm1_cnt size do not match (0x%x vs 0x%x).", width, fadt->x_pm1a_cnt_blk.register_bit_width);
+			fwts_failed_medium(fw, "32 and 64 bit versions of FADT pm1_cnt size do not match (0x%x vs 0x%x).",
+				width, fadt->x_pm1a_cnt_blk.register_bit_width);
 
 		port = fadt->x_pm1a_cnt_blk.address;	
 		width = fadt->x_pm1a_cnt_blk.register_bit_width;
